+++ docs/source/whatsnew.rst
@@ -5,6 +5,9 @@ What's New
 
 v2.3.0
 ------
+- BUG FIX: **wxcadm** no longer gets Webex licenses on Org initialization.
+- In addition to the item below, the defaults for Org creation have been changed so that data is not retrieved automatically *unless there is only one org*.
+- To handle admins who manage a large number of Orgs, the ``get_location_data`` args has been added to the :py:class:`Webex` class initialization args. This prevents **wxcadm** from making dozens of potentially unnecessary API calls and speed processing time.
 - Added Voice Message methods to the :py:class:`Me` class
 - :meth:`Webex.me` returns a :py:class:`Me` instance, which is a child of the Person instance, which has some unique methods.
 - :meth:`Webex.get_person_by_id` to search across Orgs